20160413-1655
Changed the graphics of the water.  Changed things in xls to reflect new
colour scheme ideas.  I'm not sure I want the button.  It can't be brown
though because that will be impassable, as I'm trying to make colours
dictate what I can and can't pass over.  So a closed door is impassible
(and brown).  When it's grey you can go through.  Buttons should be gray
then actually, which would work.