20160426-1247
In order to eventually power lasers, I updated the xls to include some
memory locations I'll use to track where powered fuseboxes need to be.
I also eliminated X and Y and added XY, which is relative to SC as
opposed to an X/Y coordinate on the screen.  I think it will make things
easier overall (and maybe be faster, although that's not a concern yet)