20160428-1354
While deciding how to make the laser stuff faster I decided to do some
bitwise comparisons, which meant that redesigning my alphabet was
necessary to make it work.  I've updated the alphabet tab in the XLS but
not the graphics tab, nor redefined the characters in the tbxl.

Once that's done I'll write some "on X&3 gosub/goto" lines, which should
make it faster to do all my ifs.

I'll carefully try to do that everywhere, when possible.