20160504-1650
In two lines of code I now have the Atari playing the first 4 bars of
Jump.  Each song would be ~80 characters.  At this point I don't know if
I should loop through faster and faster or not, but I think that's the
key.  A 2nd song would be cool too, but I have no idea what to choose
yet.