20160505-1515
I've re-written it a little to try to better space out the notes.
They're not much more spaced out but that's also a side effect of the
music I've chosen for now.

ctr10 is a 10 line version of what I had written already, to see how I
was doing against an 80 char limit.  I was using 8 lines.

I've also decided to start naming files using acronyms, to make typing
them in quicker when I'm loading them inside an emulator