20160525-1115
Fire is almost a functioning game at this point.  The only logic still
required is defining when a victim fails to be bounced along, scoring,
and limiting how many of victims appear.  The latter will be one a
second, then 2 a second, etc. the longer you play.  Maybe every 5-10
seconds I'll increase the frequency.  That will come from play testing.