20160526-1517
First step of the re-write - eliminate launch page.  I'm not a fan.
Then I had to replace a bunch of the "IFs" with proper TurboBasic syntax
(basically if the IF/THEN has any statements trailing it within the same
line you need to use ENDIF)

I tried replacing some variable names with ones I am more familiar with,
but it broke.  So I'll just stick with John's stuff.