20160531-1340
For Fire I'm trying to get it to limit how many get sent a second.  Or
that they can't come quicker than that anyway.

For Speed Demon i'm trying to get it to run.  I suspect that the listing
to the printer messes up some of the code, so it won't directly compile
again.  I need to figure out how to fix it before I modify it.