20160608
SPEED - I created an atr to upload to AtariAge with a finished version
of the original game, just for shits and giggles

ARENA - my next antic project.  This is the initial code with nothing
but extra spacing and a modified header

MISC - I downloaded a few basic files to take a look at.  If Arena
proves to be boring I might tackle one of these