20160609-1332
ARENA - started looking at it but it involves machine language calls.
Not sure I want to bother with that yet

MAZE - this is the original code except I tried to replace the inverse
text with \HEX, to make it more readable-ish.  I also commented on what
the text is in plain english, for clarity