20160628-1605
MINER - this is the original C64 code.  I actually believe I'm going to
be able to convert this to the XL :)

PICROSS - the beginnings of what I think is a cool idea for a 10 line
program.

TEXTTOHEX - my program to convert text values to the hex ones the
basicparser uses.