20170330-1411
Almost there.  I need to change the collision detection, so that if you
move the coffee cup directly underneath a note, you score.  I also want
to speed it up a bit and see if I can figure out why it slows down.
Maybe there's a better way to make things work in terms of timing.